% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_summary_stats}
\alias{create_summary_stats}
\title{calculate summary statistics}
\usage{
create_summary_stats(df_i)
}
\arguments{
\item{df_i}{}
}
\value{
dataframe with 1 row and summary stats
}
\description{
requires input to have initial and revision col
}
