% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vintages.R
\docType{data}
\name{gdp_us}
\alias{gdp_us}
\alias{gdp_uk}
\title{Vintages Data}
\usage{
gdp_us

gdp_uk
}
\description{
A collection of real-time datasets.
}
\details{
\itemize{
\item \link{gdp_us}: Quarterly Vintages (Billions of real dollars, seasonally adjusted)
}
}
\section{Datasets}{

\itemize{
\item \link{gdp_us}:
\itemize{
\item Source: \href{https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/real-time-data/data-files/xlsx/routputqvqd.xlsx?la=en&hash=33BCFABC01F5579AECBBFC39447EC6C9}{Federal Reserve Bank Philadelphia}
\item Timeframe: Q1 1970 - Q2 2024
\item Real-Time Vintages: January 1980 - October 2024
\item Frequency: Quarterly
\item Format: A tibble with quarterly observations and 3 variables:
\itemize{
\item \code{time}: Date of the observation.
\item \code{pub_date}: Publication date of the vintage
\item \code{values}: Numeric, real GDP (seasonally adjusted).
}
}
\item \link{gdp_uk}:
\itemize{
\item Source: \href{https://www.bankofengland.co.uk/statistics/gdp-real-time-database}{Bank of England}
\item Timeframe: Q1 1955 - Q2 2016
\item Real-Time Vintages: March 1976 - September 2016
\item Frequency: Quarterly
\item Format: A tibble with quarterly observations and 3 variables:
\itemize{
\item \code{time}: Date of the observation.
\item \code{pub_date}: Publication date of the vintage
\item \code{values}: Numeric, real GDP (seasonally adjusted).
}
}
}
}

\examples{
# Load [gdp_us]
data(gdp_us)
}
\keyword{datasets}
