% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_days_to_release}
\alias{get_days_to_release}
\title{get the number of days between first release and period end:}
\usage{
get_days_to_release(df)
}
\description{
get the number of days between first release and period end:
}
