% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_final_revision_series}
\alias{get_final_revision_series}
\title{get final revision series}
\usage{
get_final_revision_series(df, col_rel_date, col_val_date)
}
\arguments{
\item{df}{dataframe with revisions, longformat}

\item{col_rel_date}{name of column with release dates (date format)}

\item{col_val_date}{name of column with observation dates (date format)}

\item{p_month}{number of months to be added to release until "final" revision}
}
\value{
filtered df only with first release + p_month
}
\description{
get final revision series
}
