% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_fixed_release}
\alias{get_fixed_release}
\title{Extract Vintage Values from a Data Frame}
\usage{
get_fixed_release(df, years, month = NULL, quarter = NULL)
}
\arguments{
\item{df}{A data frame containing columns \code{pub_date} (publication date) and \code{ time} (observation date).}

\item{years}{The integer number of unrestricted years after \code{pub_date} for
which the values should be extracted.}

\item{month}{An optional parameter specifying the target month as a name
("July") or an integer (7). Cannot be used with \code{quarter}.}

\item{quarter}{An optional parameter specifying the target quarter (1-4).
Cannot be used with \code{month}.}
}
\value{
A filtered data frame containing values matching the
specified criteria.
}
\description{
Some statistical agencies make a final revision of their data after a certain
period of time in a give month in the year. This function extracts values
from a given month or quarter a specified  number of years after the
initial release.
}
\examples{
df <- dplyr::filter(reviser::gdp, id=="US")
dta <- get_fixed_release(df, month = "July", years = 3)
dta <- get_fixed_release(df, month = 7, years = 3)
dta <- get_fixed_release(df, quarter = 3, years = 3)

}
\seealso{
Other revision utilities: 
\code{\link{get_days_to_release}()},
\code{\link{get_first_release}()},
\code{\link{get_latest_release}()},
\code{\link{get_nth_release}()},
\code{\link{get_releases_by_date}()},
\code{\link{get_revisions}()}
}
\concept{revision utilities}
