% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_latest_release}
\alias{get_latest_release}
\title{get latest vintage}
\usage{
get_latest_release(df)
}
\arguments{
\item{df}{}

\item{col_rel_date}{}

\item{col_val_date}{}
}
\value{
filtered df only with latest data vintage
}
\description{
get latest vintage
}
