% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_latest_vintage_series}
\alias{get_latest_vintage_series}
\title{get latest vintage}
\usage{
get_latest_vintage_series(df, col_rel_date, col_val_date)
}
\arguments{
\item{df}{}

\item{col_rel_date}{}

\item{col_val_date}{}
}
\value{
filtered df only with latest data vintage
}
\description{
get latest vintage
}
