% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_realtime_series}
\alias{get_realtime_series}
\title{get first release series ("realtime")}
\usage{
get_realtime_series(df, col_rel_date, col_val_date)
}
\arguments{
\item{df}{dataframe with revisions, longformat}

\item{col_rel_date}{name of column with release dates (date format)}

\item{col_val_date}{name of column with observation dates (date format)}
}
\value{
filtered df with every first release for obs col_val_date
}
\description{
get first release series ("realtime")
}
