% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{get_revision}
\alias{get_revision}
\title{Title}
\usage{
get_revision(df, col_rel_date, col_val_date, i)
}
\arguments{
\item{df}{}

\item{col_rel_date}{}

\item{col_val_date}{}

\item{i}{}
}
\description{
Title
}
