% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{KalmanNowcastPrecision}
\alias{KalmanNowcastPrecision}
\title{Univariate simple Kalman filter}
\usage{
KalmanNowcastPrecision(x, y, m0, C0, initial_params)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
MLE of state space model - Kalman filter
}
\description{
Univariate simple Kalman filter
}
