% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{kk_matrices}
\alias{kk_matrices}
\title{Create matrices to estimate Kishor-Koenig model}
\usage{
kk_matrices(e, model, params = NULL, type = "numeric")
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
MLE of state space model - Kalman filter and smoother
}
\description{
Create matrices to estimate Kishor-Koenig model
}
