% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{MincerZarnowitz}
\alias{MincerZarnowitz}
\title{Title}
\usage{
MincerZarnowitz(df_full, df_final, significance)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\description{
Title
}
