% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{nobs.jvn_model}
\alias{nobs.jvn_model}
\title{Number of observations used to fit a JVN model}
\usage{
\method{nobs}{jvn_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{jvn_model}.}

\item{...}{Ignored.}
}
\value{
A single integer.
}
\description{
Returns the effective number of observations behind the reported
information criteria. Under the default \code{ic_n = "Tp"} this is the number
of time periods times the number of vintages modeled.
}
\examples{
\donttest{
gdp_growth <- dplyr::filter(
  tsbox::ts_pc(reviser::gdp),
  id == "EA",
  time >= min(pub_date),
  time <= as.Date("2020-01-01")
)
gdp_growth <- tidyr::drop_na(gdp_growth)
df <- get_nth_release(gdp_growth, n = 0:3)

fit <- jvn_nowcast(df = df, e = 4, ar_order = 2, include_noise = FALSE)
nobs(fit)
}
}
\seealso{
Other revision nowcasting: 
\code{\link{coef.jvn_model}()},
\code{\link{coef.kk_model}()},
\code{\link{fitted.jvn_model}()},
\code{\link{fitted.kk_model}()},
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{logLik.jvn_model}()},
\code{\link{logLik.kk_model}()},
\code{\link{nobs.kk_model}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{predict.jvn_model}()},
\code{\link{predict.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{print.kk_model}()},
\code{\link{residuals.jvn_model}()},
\code{\link{residuals.kk_model}()},
\code{\link{states}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()},
\code{\link{vcov.jvn_model}()},
\code{\link{vcov.kk_model}()}
}
\concept{revision nowcasting}
