% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{nobs.kk_model}
\alias{nobs.kk_model}
\title{Number of observations used to fit a KK model}
\usage{
\method{nobs}{kk_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{kk_model}.}

\item{...}{Ignored.}
}
\value{
A single integer.
}
\description{
Returns the effective number of observations behind the reported
information criteria. Under the default \code{ic_n = "Tp"} this is the number
of time periods times the number of releases modeled.
}
\examples{
df <- get_nth_release(
  tsbox::ts_span(
    tsbox::ts_pc(dplyr::filter(reviser::gdp, id == "US")),
    start = "1980-01-01"
  ),
  n = 0:1
)
df <- na.omit(dplyr::select(df, -c("id", "pub_date")))
fit <- kk_nowcast(df, e = 1, model = "KK", method = "MLE")
nobs(fit)
}
\seealso{
Other revision nowcasting: 
\code{\link{coef.jvn_model}()},
\code{\link{coef.kk_model}()},
\code{\link{fitted.jvn_model}()},
\code{\link{fitted.kk_model}()},
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{logLik.jvn_model}()},
\code{\link{logLik.kk_model}()},
\code{\link{nobs.jvn_model}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{predict.jvn_model}()},
\code{\link{predict.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{print.kk_model}()},
\code{\link{residuals.jvn_model}()},
\code{\link{residuals.kk_model}()},
\code{\link{states}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()},
\code{\link{vcov.jvn_model}()},
\code{\link{vcov.kk_model}()}
}
\concept{revision nowcasting}
