% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{nobs.revision_model}
\alias{nobs.revision_model}
\title{Number of Observations Used to Fit a Revision Model}
\usage{
\method{nobs}{revision_model}(object, ...)
}
\arguments{
\item{object}{A fitted model object inheriting from \link{revision_model}, such
as a \code{kk_model} or a \code{jvn_model}.}

\item{...}{Ignored.}
}
\value{
A single integer.
}
\description{
Returns the effective number of observations behind the reported
information criteria. Under the default \code{ic_n = "Tp"} this is the number
of time periods times the number of releases (KK) or vintages (JVN)
modeled.
}
\examples{
df <- get_nth_release(
  tsbox::ts_span(
    tsbox::ts_pc(dplyr::filter(reviser::gdp, id == "US")),
    start = "1980-01-01"
  ),
  n = 0:1
)
df <- na.omit(dplyr::select(df, -c("id", "pub_date")))
fit <- kk_nowcast(df, e = 1, model = "KK", method = "MLE")
nobs(fit)
}
\seealso{
Other revision nowcasting: 
\code{\link{coef.revision_model}()},
\code{\link{fitted.revision_model}()},
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{logLik.revision_model}()},
\code{\link{plot.revision_model}()},
\code{\link{predict.revision_model}()},
\code{\link{print.revision_model}()},
\code{\link{residuals.revision_model}()},
\code{\link{revision_model}},
\code{\link{states}()},
\code{\link{summary.revision_model}()},
\code{\link{vcov.revision_model}()}
}
\concept{revision nowcasting}
