% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvn.R
\name{plot.jvn_model}
\alias{plot.jvn_model}
\title{Plot JVN model results}
\usage{
\method{plot}{jvn_model}(x, state = "true_lag_0", type = "filtered", ...)
}
\arguments{
\item{x}{An object of class \code{jvn_model}.}

\item{state}{Character scalar giving the state to visualize. Defaults to
\code{"true_lag_0"}.}

\item{type}{Character scalar indicating whether \code{"filtered"} or \code{"smoothed"}
estimates should be plotted.}

\item{...}{Additional arguments passed to \code{plot.revision_model()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot filtered or smoothed estimates for a selected state from a fitted
\code{jvn_model}.
}
\details{
This method requires \code{x$states} to be available. If the model was
fitted with \code{solver_options$return_states = FALSE}, plotting is not
possible.
}
\examples{
\donttest{
gdp_growth <- dplyr::filter(
  tsbox::ts_pc(reviser::gdp),
  id == "EA",
  time >= min(pub_date),
  time <= as.Date("2020-01-01")
)
gdp_growth <- tidyr::drop_na(gdp_growth)
df <- get_nth_release(gdp_growth, n = 0:3)

result <- jvn_nowcast(
  df = df,
  e = 4,
  ar_order = 2,
  h = 0,
  include_news = TRUE,
  include_noise = TRUE
)
plot(result)
}

}
\seealso{
Other revision nowcasting: 
\code{\link{coef.jvn_model}()},
\code{\link{coef.kk_model}()},
\code{\link{fitted.jvn_model}()},
\code{\link{fitted.kk_model}()},
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{logLik.jvn_model}()},
\code{\link{logLik.kk_model}()},
\code{\link{nobs.jvn_model}()},
\code{\link{nobs.kk_model}()},
\code{\link{plot.kk_model}()},
\code{\link{predict.jvn_model}()},
\code{\link{predict.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{print.kk_model}()},
\code{\link{residuals.jvn_model}()},
\code{\link{residuals.kk_model}()},
\code{\link{states}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()},
\code{\link{vcov.jvn_model}()},
\code{\link{vcov.kk_model}()}
}
\concept{revision nowcasting}
