% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot.tbl_pubdate}
\alias{plot.tbl_pubdate}
\title{Plot Method for Publication Date Vintages}
\usage{
\method{plot}{tbl_pubdate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_pubdate}.}

\item{...}{Additional arguments passed to plot_vintages.}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for Publication Date Vintages
}
\seealso{
Other revision graphs: 
\code{\link{plot.tbl_release}()},
\code{\link{plot_vintages}()},
\code{\link{theme_reviser}()}
}
\concept{revision graphs}
