% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot.tbl_release}
\alias{plot.tbl_release}
\title{Plot Method for Release Vintages}
\usage{
\method{plot}{tbl_release}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_release}.}

\item{...}{Additional arguments passed to plot_vintages.}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for Release Vintages
}
\seealso{
Other revision graphs: 
\code{\link{plot.tbl_pubdate}()},
\code{\link{plot_vintages}()},
\code{\link{theme_reviser}()}
}
\concept{revision graphs}
