% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage-methods.R
\name{plot.tbl_vintage}
\alias{plot.tbl_vintage}
\title{Plot Method for Vintages Data}
\usage{
\method{plot}{tbl_vintage}(x, ...)
}
\arguments{
\item{x}{An object inheriting from \link{tbl_vintage}, such as a \code{tbl_pubdate}
or a \code{tbl_release}.}

\item{...}{Additional arguments passed to \code{\link[=plot_vintages]{plot_vintages()}}.}
}
\value{
A ggplot2 object.
}
\description{
Plots a vintages object along the dimension implied by its class:
publication date for a \code{tbl_pubdate}, release number for a \code{tbl_release}.
Defined once for the parent class \link{tbl_vintage} and inherited by both.
\code{\link[=plot_vintages]{plot_vintages()}} remains the entry point when the dimension, title or
plot type need to be set explicitly.
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
plot(df)

release_data <- get_nth_release(df, n = 0:5)
plot(release_data)
}
\seealso{
Other revision graphs: 
\code{\link{plot_vintages}()},
\code{\link{theme_reviser}()}
}
\concept{revision graphs}
