% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{print.kk_model}
\alias{print.kk_model}
\title{Print Method for KK Model}
\usage{
\method{print}{kk_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{kk_model}.}

\item{...}{Additional arguments passed to \code{summary.kk_model}.}
}
\value{
The function returns the input \code{x} invisibly.
}
\description{
Default print method for \code{kk_model} objects.
Wraps the \code{summary} method for a consistent output.
}
\examples{
df <- get_nth_release(
  tsbox::ts_span(
    tsbox::ts_pc(
      dplyr::filter(reviser::gdp, id == "US")
    ),
    start = "1980-01-01"
  ),
  n = 0:1
)
df <- dplyr::select(df, -c("id", "pub_date"))
df <- na.omit(df)

e <- 1
h <- 2
result <- kk_nowcast(df, e, h = h, model = "Kishor-Koenig", method = "MLE")
result

}
\seealso{
Other revision nowcasting: 
\code{\link{coef.jvn_model}()},
\code{\link{coef.kk_model}()},
\code{\link{fitted.jvn_model}()},
\code{\link{fitted.kk_model}()},
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{logLik.jvn_model}()},
\code{\link{logLik.kk_model}()},
\code{\link{nobs.jvn_model}()},
\code{\link{nobs.kk_model}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{predict.jvn_model}()},
\code{\link{predict.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{residuals.jvn_model}()},
\code{\link{residuals.kk_model}()},
\code{\link{states}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()},
\code{\link{vcov.jvn_model}()},
\code{\link{vcov.kk_model}()}
}
\concept{revision nowcasting}
