% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{print.lst_efficient}
\alias{print.lst_efficient}
\title{Print Method for Efficient Release Results}
\usage{
\method{print}{lst_efficient}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lst_efficient}.}

\item{...}{Additional arguments (not used).}
}
\value{
The function returns the input \code{x} invisibly.
}
\description{
Print Method for Efficient Release Results
}
\seealso{
Other revision analysis: 
\code{\link{diagnose}()},
\code{\link{diagnose.revision_summary}()},
\code{\link{get_first_efficient_release}()},
\code{\link{get_revision_analysis}()},
\code{\link{print.revision_summary}()},
\code{\link{summary.lst_efficient}()},
\code{\link{summary.revision_summary}()}
}
\concept{revision analysis}
