% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage-methods.R
\name{print.tbl_vintage}
\alias{print.tbl_vintage}
\title{Print Method for Vintages Data}
\usage{
\method{print}{tbl_vintage}(x, ...)
}
\arguments{
\item{x}{An object inheriting from \link{tbl_vintage}, such as a \code{tbl_pubdate}
or a \code{tbl_release}.}

\item{...}{Additional arguments passed to the next print method.}
}
\value{
The input \code{x} is returned invisibly.
}
\description{
Print method for objects inheriting from \link{tbl_vintage}. Delegates to the
tibble print method, which calls \code{\link[=tbl_sum.tbl_vintage]{tbl_sum.tbl_vintage()}} to generate the
custom header.
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
release_data <- get_nth_release(df, n = 0:3)
print(release_data)
}
\seealso{
Other helpers: 
\code{\link{summary.tbl_vintage}()},
\code{\link{tbl_sum.tbl_vintage}()},
\code{\link{tbl_vintage}},
\code{\link{validate_vintages}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
