% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{ss_to_kk}
\alias{ss_to_kk}
\title{Cast state space matrices back into Generalized Kishor-Koenig form}
\usage{
ss_to_kk(Z, Tmat, V, W)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
MLE of state space model - Kalman filter and smoother
}
\description{
Cast state space matrices back into Generalized Kishor-Koenig form
}
