% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{summary.kk_model}
\alias{summary.kk_model}
\title{Summarize the results of a kk_model object.}
\usage{
\method{summary}{kk_model}(object, ...)
}
\arguments{
\item{object}{A list of class 'kk_model' produced by the
\code{\link{kk_nowcast}} function.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing two data frames:
\item{final_release_metrics}{A data frame with MSE, RMSE, and MAE against
the final release.}
\item{true_efficient_release_metrics}{A data frame with MSE, RMSE, and
MAE against the true efficient release.}
}
\description{
This function calculates and prints the Mean Squared Error (MSE), Root Mean
Squared Error (RMSE), and Mean Absolute Error (MAE) of the filtered state
variables against both the final release and the true efficient release.
}
\examples{
# Assuming 'kk_model_obj' is the result of kk_nowcast(your_data, ...)
# and 'your_data' is the original data frame.
# results <- summary.kk_model(kk_model_obj, your_data)

}
\seealso{
Other revision nowcasting: 
\code{\link{kk_nowcast}()}
}
\concept{revision nowcasting}
