% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{summary.list_eff_rel}
\alias{summary.list_eff_rel}
\title{summary of efficient release models}
\usage{
\method{summary}{list_eff_rel}(object, ...)
}
\arguments{
\item{output}{from get_first_efficient_release}
}
\description{
summary of efficient release models
}
