% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.tbl_release}
\alias{summary.tbl_release}
\title{Summary Method for Release Vintages}
\usage{
\method{summary}{tbl_release}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tbl_release}.}

\item{...}{Additional arguments (not used).}
}
\value{
The function returns a summary tibble invisibly.
}
\description{
Summary Method for Release Vintages
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
# Long format
release_data <- get_nth_release(df, n = 0:3)
summary(release_data)

# Wide format
wide_release <- vintages_wide(release_data, names_from = "release")
summary(wide_release$US)
}
\seealso{
Other helpers: 
\code{\link{print.tbl_pubdate}()},
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{tbl_sum.tbl_pubdate}()},
\code{\link{tbl_sum.tbl_release}()},
\code{\link{validate_vintages}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
