% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vintages_rename}
\alias{vintages_rename}
\title{Function to rename columns in accordance with the package}
\usage{
vintages_rename(
  df,
  col_time = NULL,
  col_pub_date = NULL,
  col_value = NULL,
  col_id = NULL
)
}
\arguments{
\item{df}{data.frame or tibble}

\item{col_time}{name of column with observation dates (date format)}

\item{col_pub_date}{name of column with release dates (date format)}

\item{col_value}{name of column with values}

\item{col_id}{name of column with id}
}
\description{
Function to rename columns in accordance with the package
}
