% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vintages.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Vintages Data}
\usage{
gdp
}
\description{
A collection of datasets containing economic indicators for Switzerland.
}
\details{
\itemize{
\item \link{gdp}: Quarterly Vintages (Billions of real dollars, seasonally adjusted)
}
}
\section{Datasets}{

\itemize{
\item \link{gdp}:
\itemize{
\item Source: \href{https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/real-time-data/data-files/xlsx/routputqvqd.xlsx?la=en&hash=33BCFABC01F5579AECBBFC39447EC6C9}{Federal Reserve Bank Philadelphia}
\item Timeframe: Q1 1970 - Q2 2024
\item Real-Time Vintage: January 1980 - October 2024
\item Frequency: Quarterly
\item Format: A tibble with quarterly observations and 2 variables:
\itemize{
\item \code{time}: Date, the quarter and year of the observation.
\item \code{values}: Numeric, the real GDP (seasonally adjusted).
}
}
}
}

\examples{
# Load and plot [gdp]
data(gdp)
}
\keyword{datasets}
