% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vintages.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Vintages Data}
\usage{
gdp
}
\description{
A collection of real-time datasets.
}
\details{
\itemize{
\item \link{gdp}: Quarterly Vintages (Billions of real dollars, seasonally adjusted)
\item Timeframe: Q1 1980 - Q4 2022
\item Real-Time Vintages: Q1 2003 - Q4 2022
\item Frequency: Quarterly
\item Format: A tibble with quarterly observations and 3 variables:
\itemize{
\item \code{time}: Date of the observation.
\item \code{pub_date}: Publication date of the vintage
\item \code{values}: Numeric, real GDP (seasonally adjusted).
\item \code{id}: Country code
}
}
}
\section{Sources}{

\itemize{
\item All the data is taken from the realtime database of Indergand and Leist (2014).
\item Countries:
\item CHE:
\itemize{
\item Switzerland
\item Source: \href{https://www.seco.admin.ch/seco/de/home/wirtschaftslage---wirtschaftspolitik/Wirtschaftslage/bip-quartalsschaetzungen-/daten.html}{SECO}
}
\item US:
\itemize{
\item United States
\item Sources: FRED, OECD
}
\item EA:
\itemize{
\item Euro Area
\item Sources: Eurostat, OECD
}
\item JP:
\itemize{
\item Japan
\item Sources: Cabinet Office (Japan), OECD
}
}
}

\examples{
# Load [gdp] dataset
data(gdp)
}
\references{
Indergand, R., Leist, S. A Real-Time Data Set for Switzerland. Swiss J Economics Statistics 150, 331–352 (2014). https://doi.org/10.1007/BF03399410
}
\keyword{datasets}
