% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vintages.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Vintages Data}
\format{
A tibble with quarterly observations and 4 variables:
\describe{
\item{time}{Date of the observation}
\item{pub_date}{Publication date of the vintage}
\item{value}{Numeric, real GDP (seasonally adjusted)}
\item{id}{Country code}
}
}
\usage{
gdp
}
\description{
A collection of real-time datasets.
}
\details{
\itemize{
\item GDP: Quarterly Vintages (Billions of real dollars, seasonally adjusted)
\item Timeframe: Q1 1980 - Q4 2024
\item Real-Time Vintages: Q4 2002 - Q4 2024
}
}
\section{Sources}{

\itemize{
\item All the data is from the realtime database of Indergand and Leist (2014).
\strong{Countries}:
\item CHE:
\itemize{
\item Switzerland
\item Source: SECO
}
\item US:
\itemize{
\item United States
\item Sources: FRED, OECD
}
\item EA:
\itemize{
\item Euro Area
\item Sources: Eurostat, OECD
}
\item JP:
\itemize{
\item Japan
\item Sources: Cabinet Office (Japan), OECD
}
}
}

\examples{
# Load gdp dataset
data(gdp)
head(gdp)
}
\references{
Indergand, R., Leist, S. A Real-Time Data Set for Switzerland.
Swiss J Economics Statistics 150, 331–352 (2014).
\doi{10.1007/BF03399410}
}
\concept{dataset}
\keyword{datasets}
