% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_first_efficient_release}
\alias{get_first_efficient_release}
\title{get first efficient release using Mincer Zarnowitz regression}
\usage{
get_first_efficient_release(df, final_release, significance = 0.05)
}
\arguments{
\item{df}{}
}
\value{
filtered df only with first data vintage
}
\description{
get first efficient release using Mincer Zarnowitz regression
}
