% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_first_release}
\alias{get_first_release}
\title{get first vintage}
\usage{
get_first_release(df)
}
\arguments{
\item{df}{}
}
\value{
filtered df only with first data vintage
}
\description{
get first vintage
}
