% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_latest_release}
\alias{get_latest_release}
\title{Extract the Latest Data Release (Vintage)}
\usage{
get_latest_release(df)
}
\arguments{
\item{df}{A data frame containing data vintages. The data frame must include
the columns \code{pub_date} (publication date of the release) and
\code{time} (the corresponding time period for the data).}
}
\value{
A filtered data frame containing only the most recent release(s).
The resulting data frame is assigned the class \code{tbl_release} to indicate
its structure.
}
\description{
Filters the input dataset to return the most recent release (or vintage)
for each time period.
}
\details{
For each time period, the function identifies the release with the latest
publication date (\code{pub_date})  and adds a column \code{release} that labels the
release as \code{release_N}, where \code{N} is the release index (zero indexed).
}
\examples{
# Example data
df <- dplyr::filter(reviser::gdp, id=="US")

# Get the latest release for each time period
latest_release <- get_latest_release(df)

}
