% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_nth_release}
\alias{get_nth_release}
\title{get nth vintage}
\usage{
get_nth_release(df, n = 1)
}
\arguments{
\item{df}{}
}
\value{
filtered df only with first data vintage
}
\description{
get nth vintage
}
