% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{kk_nowcast}
\alias{kk_nowcast}
\title{Generalized Kishor-Koenig model nowcasting}
\usage{
kk_nowcast(
  df,
  e,
  h = 1,
  model = "Kishor-Koenig",
  estim = "MLE",
  start_mat = 0.4,
  start_cov = 0.4,
  m0 = 0,
  C0 = 1e-06,
  trace = 0
)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
MLE of state space model - Kalman filter and smoother
}
\description{
Generalized Kishor-Koenig model nowcasting
}
