% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{kk_to_ss}
\alias{kk_to_ss}
\title{Cast Generalized Kishor-Koenig matrices into state space form}
\usage{
kk_to_ss(II, FF, GG, R, H, epsilon = 1e-06)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
MLE of state space model - Kalman filter and smoother
}
\description{
Cast Generalized Kishor-Koenig matrices into state space form
}
