% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{MincerZarnowitz}
\alias{MincerZarnowitz}
\title{Mincer-Zarnowitz regression}
\usage{
MincerZarnowitz(df_full, df_final, significance)
}
\arguments{
\item{df_full}{}

\item{df_final}{}

\item{significance}{}
}
\value{
Statistics from MZ-regression
}
\description{
Mincer-Zarnowitz regression
}
