% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvn.R
\name{plot.jvn_model}
\alias{plot.jvn_model}
\title{Plot JVN Model Results}
\usage{
\method{plot}{jvn_model}(x, state = "true_lag_0", type = "filtered", ...)
}
\arguments{
\item{x}{An object of class 'jvn_model'}

\item{state}{String. The name of the state to visualize (e.g., "state_1").}

\item{type}{String. Type of estimate to plot: "filtered" or "smoothed".}

\item{...}{Additional arguments passed to theme_reviser.}
}
\value{
A ggplot2 object visualizing the specified state estimates.
}
\description{
Plot JVN Model Results
}
