% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{plot.kk_model}
\alias{plot.kk_model}
\title{Plot JVN Model Results}
\usage{
\method{plot}{kk_model}(x, state = NULL, type = "filtered", ...)
}
\arguments{
\item{x}{An object of class 'jvn_model'}

\item{state}{String. The name of the state to visualize (e.g., "state_1").}

\item{type}{String. Type of estimate to plot: "filtered" or "smoothed".}

\item{...}{Additional arguments passed to theme_reviser.}
}
\value{
A ggplot2 object visualizing the specified state estimates.
}
\description{
Plot JVN Model Results
}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{print.jvn_model}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()}
}
\concept{revision nowcasting}
