% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{plot.kk_model}
\alias{plot.kk_model}
\title{Plot Kishor-Koenig Model Results}
\usage{
\method{plot}{kk_model}(x, state = NULL, type = "filtered", ...)
}
\arguments{
\item{x}{An object of class \code{kk_model}.}

\item{state}{Character scalar giving the state to visualize. If \code{NULL}, the
first available state is used.}

\item{type}{Character scalar indicating whether \code{"filtered"} or \code{"smoothed"}
estimates should be plotted.}

\item{...}{Additional arguments passed to \code{plot.revision_model()}.}
}
\value{
A \code{ggplot2} object visualizing the specified state estimates.
}
\description{
Plot filtered or smoothed estimates for a selected state from a fitted
\code{kk_model}.
}
\details{
This method requires \code{x$states} to be available. If the model was
fitted with \code{solver_options$return_states = FALSE}, plotting is not
possible.
}
\examples{
df <- get_nth_release(
  tsbox::ts_span(
    tsbox::ts_pc(
      dplyr::filter(reviser::gdp, id == "US")
    ),
    start = "1980-01-01"
  ),
  n = 0:1
)
df <- dplyr::select(df, -c("id", "pub_date"))
df <- na.omit(df)

e <- 1 # Number of efficient release
h <- 2 # Forecast horizon
result <- kk_nowcast(df, e, h = h, model = "Kishor-Koenig")

plot(result)

}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{print.jvn_model}()},
\code{\link{print.kk_model}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()}
}
\concept{revision nowcasting}
