% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_vintages}
\alias{plot_vintages}
\alias{theme_reviser}
\alias{colors_reviser}
\alias{scale_color_reviser}
\alias{scale_fill_reviser}
\title{Functions to plot vintages data}
\usage{
plot_vintages(
  df,
  type = "line",
  dim_col = "pub_date",
  title = "",
  subtitle = "",
  ylab = ""
)

theme_reviser(
  base_size = 12,
  legend.position = "bottom",
  legend.direction = "horizontal"
)

colors_reviser()

scale_color_reviser(...)

scale_fill_reviser(...)
}
\description{
Functions to plot vintages data
}
