% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvn.R
\name{print.jvn_model}
\alias{print.jvn_model}
\title{Print Method for JVN Model}
\usage{
\method{print}{jvn_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{jvn_model}.}

\item{...}{Additional arguments passed to \code{summary.jvn_model}.}
}
\value{
The function returns the input \code{x} invisibly.
}
\description{
Default print method for \code{jvn_model} objects.
Wraps the \code{summary} method for a consistent output.
}
\examples{
gdp_growth <- dplyr::filter(
  tsbox::ts_pc(
    reviser::gdp
  ), id \%in\% c("EA"),
  time >= min(pub_date),
  time <= as.Date("2020-01-01")
)
gdp_growth <- tidyr::drop_na(gdp_growth)
df <- get_nth_release(gdp_growth, n = 0:4)

# Estimate model
result <- jvn_nowcast(
  df = df,
  e = 3,
  ar_order = 2,
  h = 4,
  include_news = TRUE,
  include_noise = TRUE
)
result

}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{print.kk_model}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()}
}
\concept{revision nowcasting}
