% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvn.R
\name{print.jvn_model}
\alias{print.jvn_model}
\title{Print method for JVN model objects}
\usage{
\method{print}{jvn_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{jvn_model}.}

\item{...}{Additional arguments passed to \code{summary.jvn_model()}.}
}
\value{
The input object, invisibly.
}
\description{
Default print method for \code{jvn_model} objects. This method dispatches to
\code{summary.jvn_model()} for a consistent console display.
}
\examples{
\donttest{
gdp_growth <- dplyr::filter(
  tsbox::ts_pc(reviser::gdp),
  id == "EA",
  time >= min(pub_date),
  time <= as.Date("2020-01-01")
)
gdp_growth <- tidyr::drop_na(gdp_growth)
df <- get_nth_release(gdp_growth, n = 0:3)

result <- jvn_nowcast(
  df = df,
  e = 4,
  ar_order = 2,
  h = 0,
  include_news = TRUE,
  include_noise = TRUE
)
result
}

}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{print.kk_model}()},
\code{\link{summary.jvn_model}()},
\code{\link{summary.kk_model}()}
}
\concept{revision nowcasting}
