% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tbl_pubdate}
\alias{print.tbl_pubdate}
\title{Print Method for Publication Date Vintages}
\usage{
\method{print}{tbl_pubdate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_pubdate}.}

\item{...}{Additional arguments passed to print.}
}
\value{
The function returns the input \code{x} invisibly.
}
\description{
Print Method for Publication Date Vintages
}
\seealso{
Other helpers: 
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{summary.tbl_release}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
