% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tbl_pubdate}
\alias{print.tbl_pubdate}
\title{Print Method for Publication Date Vintages}
\usage{
\method{print}{tbl_pubdate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_pubdate}.}

\item{...}{Additional arguments passed to the next print method.}
}
\value{
The input \code{x} is returned invisibly.
}
\description{
Print method for objects of class \code{tbl_pubdate}. This method delegates
to the tibble print method, which will automatically call
\code{tbl_sum.tbl_pubdate}
to generate the custom header.
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
wide_data <- vintages_wide(df)
print(wide_data$US)
}
\seealso{
Other helpers: 
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{summary.tbl_release}()},
\code{\link{tbl_sum.tbl_pubdate}()},
\code{\link{tbl_sum.tbl_release}()},
\code{\link{validate_vintages}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
