% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tbl_release}
\alias{print.tbl_release}
\title{Print Method for Release Vintages}
\usage{
\method{print}{tbl_release}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_release}.}

\item{...}{Additional arguments passed to the next print method.}
}
\value{
The input \code{x} is returned invisibly.
}
\description{
Print method for objects of class \code{tbl_release}.
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
release_data <- get_nth_release(df, n = 0:3)
print(release_data)
}
\seealso{
Other helpers: 
\code{\link{print.tbl_pubdate}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{summary.tbl_release}()},
\code{\link{tbl_sum.tbl_pubdate}()},
\code{\link{tbl_sum.tbl_release}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
