% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviser-package.R
\docType{package}
\name{reviser-package}
\alias{reviser}
\alias{reviser-package}
\title{reviser: Tools for Studying Revision Properties in Real-Time Time Series Vintages}
\description{
To learn more about reviser, start with the vignettes:
\code{browseVignettes(package = "reviser")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://p-wegmueller.github.io/reviser/}
}

}
\author{
\strong{Maintainer}: Marc Burri \email{marc.burri91@gmail.com} (\href{https://orcid.org/0000-0001-8974-9090}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Philipp Wegmueller \email{philipp.wemueller@seco.admin.ch} [copyright holder]
}

}
\keyword{internal}
