% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviser-package.R
\docType{package}
\name{reviser-package}
\alias{reviser}
\alias{reviser-package}
\title{reviser: Analyzing Revisions in Real-Time Time Series Vintages}
\description{
\strong{reviser} is an R package designed for working with time-series vintages
data. The package provides tools to clean, visualize, and analyze time-series
revisions. To learn more about reviser, start with the vignettes:
\code{browseVignettes(package = "reviser")}
The package will be actively developed and maintained.

Useful links:
\itemize{
\item \url{https://docs.ropensci.org/reviser/}
\item Report bugs at \url{https://github.com/ropensci/reviser/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/reviser/}
  \item \url{https://github.com/ropensci/reviser}
  \item Report bugs at \url{https://github.com/ropensci/reviser/issues}
}

}
\author{
\strong{Maintainer}: Marc Burri \email{marc.burri91@gmail.com} (\href{https://orcid.org/0000-0001-8974-9090}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Philipp Wegmueller \email{philipp.wegmueller@seco.admin.ch} [copyright holder]
}

}
\concept{package overview}
\keyword{internal}
