% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvn.R
\name{summary.jvn_model}
\alias{summary.jvn_model}
\title{Summary Method for JVN Model}
\usage{
\method{summary}{jvn_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{jvn_model}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The function returns the input \code{object} invisibly.
}
\description{
Computes and displays a summary of the results from a
Jacobs-Van Norden (JVN) model fit, including convergence status,
information criteria, and parameter estimates.
}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{summary.kk_model}()}
}
\concept{revision nowcasting}
