% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{summary.kk_model}
\alias{summary.kk_model}
\title{Summary Method for KK Model}
\usage{
\method{summary}{kk_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{kk_model}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The function returns the input \code{object} invisibly.
}
\description{
Computes and displays a summary of the results from a
Kishor-Koenig (KK) model fit, including convergence status,
information criteria, and parameter estimates.
}
\examples{
df <- get_nth_release(
  tsbox::ts_span(
    tsbox::ts_pc(
      dplyr::filter(reviser::gdp, id == "US")
    ),
    start = "1980-01-01"
  ),
  n = 0:1
)
df <- dplyr::select(df, -c("id", "pub_date"))
df <- na.omit(df)

e <- 1
h <- 2
result <- kk_nowcast(df, e, h = h, model = "Kishor-Koenig", method = "MLE")
summary(result)

}
\seealso{
Other revision nowcasting: 
\code{\link{jvn_nowcast}()},
\code{\link{kk_nowcast}()},
\code{\link{plot.jvn_model}()},
\code{\link{plot.kk_model}()},
\code{\link{print.jvn_model}()},
\code{\link{print.kk_model}()},
\code{\link{summary.jvn_model}()}
}
\concept{revision nowcasting}
