% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{summary.lst_efficient}
\alias{summary.lst_efficient}
\title{Summary of Efficient Release Models}
\usage{
\method{summary}{lst_efficient}(object, ...)
}
\arguments{
\item{object}{An output object from the \code{get_first_efficient_release} function.
The object must be of class \code{list_eff_rel}.}

\item{...}{Additional arguments (not used).}
}
\value{
Returns a tibble with the following columns:
\itemize{
\item \code{id}: The identifier of the time series (if present in input data).
\item \code{e}: The index of the first efficient release.
\item \code{alpha}: The intercept coefficient of the regression model.
\item \code{beta}: The coefficient of the slope.
\item \code{p-value}: The p-value of the joint hypothesis test (alpha = 0 and beta = 1).
\item \code{n_tested}: The number of releases tested.
}
}
\description{
Provides a detailed summary of the regression model and hypothesis test for the first efficient release
identified by the \code{get_first_efficient_release} function.
}
\details{
This function prints the following information:
\itemize{
\item The index of the first efficient release.
\item A summary of the regression model fitted for the efficient release, which includes coefficients,
R-squared values, and other relevant statistics.
\item The hypothesis test results for the efficient release, showing the test statistic and p-value
for the null hypothesis of unbiasedness and efficiency.
}

The function assumes the object includes:
\itemize{
\item \code{e}: The index of the first efficient release (0-based).
\item \code{models}: A list of linear regression models for each release.
\item \code{tests}: A list of hypothesis test results corresponding to each release.
}
}
\examples{
# Example usage
df <- get_nth_release(tsbox::ts_pc(dplyr::filter(reviser::gdp , id=="US")), n = 1:4)

final_release <- get_nth_release(tsbox::ts_pc(dplyr::filter(reviser::gdp, id=="US")), n = 10)

# Identify the first efficient release
result <- get_first_efficient_release(df, final_release, significance = 0.05)
summary(result)

}
