% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.tbl_pubdate}
\alias{summary.tbl_pubdate}
\title{Summary Method for Publication Date Vintages}
\usage{
\method{summary}{tbl_pubdate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tbl_pubdate}.}

\item{...}{Additional arguments (not used).}
}
\value{
The function returns a summary tibble invisibly.
}
\description{
Summary Method for Publication Date Vintages
}
\examples{
df <- dplyr::filter(reviser::gdp, id == "US")
wide_data <- vintages_wide(df)
summary(wide_data$US)
}
\seealso{
Other helpers: 
\code{\link{print.tbl_pubdate}()},
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_release}()},
\code{\link{tbl_sum.tbl_pubdate}()},
\code{\link{tbl_sum.tbl_release}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
