% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.tbl_release}
\alias{summary.tbl_release}
\title{Summary Method for Release Vintages}
\usage{
\method{summary}{tbl_release}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tbl_release}.}

\item{...}{Additional arguments (not used).}
}
\value{
The function returns a summary tibble invisibly.
}
\description{
Summary Method for Release Vintages
}
\seealso{
Other helpers: 
\code{\link{print.tbl_pubdate}()},
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
