% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_sum.tbl_release}
\alias{tbl_sum.tbl_release}
\title{Tibble Summary for Release Vintages}
\usage{
\method{tbl_sum}{tbl_release}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tbl_release}.}

\item{...}{Additional arguments (unused).}
}
\value{
A named character vector where names are labels and values are
the corresponding information.
}
\description{
Provides a custom header for objects of class \code{tbl_release} when printed.
}
\seealso{
Other helpers: 
\code{\link{print.tbl_pubdate}()},
\code{\link{print.tbl_release}()},
\code{\link{summary.tbl_pubdate}()},
\code{\link{summary.tbl_release}()},
\code{\link{tbl_sum.tbl_pubdate}()},
\code{\link{vintages_long}()},
\code{\link{vintages_wide}()}
}
\concept{helpers}
