% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{theme_reviser}
\alias{theme_reviser}
\alias{colors_reviser}
\alias{scale_color_reviser}
\alias{scale_fill_reviser}
\title{Custom Visualization Theme and Color Scales for Reviser}
\usage{
theme_reviser(
  base_size = 12,
  legend.position = "bottom",
  legend.direction = "horizontal"
)

colors_reviser()

scale_color_reviser(...)

scale_fill_reviser(...)
}
\arguments{
\item{base_size}{Numeric. The base font size for the theme. Default is 12.}

\item{legend.position}{Character. Position of the legend. Default is "bottom".}

\item{legend.direction}{Character. Direction of the legend. Default is "horizontal".}

\item{...}{Additional arguments passed to the ggplot2 scale functions.}
}
\value{
A customized ggplot2 theme, color palette, or scale.
}
\description{
These functions provide a custom visualization theme and color scales for use with ggplot2, inspired by the \code{tsbox} package.
}
\details{
\itemize{
\item{\code{theme_reviser}: Defines a minimal theme with custom adjustments for axis titles, plot titles, subtitles, captions, and legend positioning.}
\item{\code{colors_reviser}: Provides a predefined set of colors, including a soft black, a palette suitable for colorblind readers, and additional colors for extended use.}
\item{\code{scale_color_reviser}: A ggplot2 color scale that uses the custom \code{colors_reviser} palette.}
\item{\code{scale_fill_reviser}: A ggplot2 fill scale that uses the custom \code{colors_reviser} palette.}
}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point(size = 3) +
  theme_reviser() +
  scale_color_reviser()

}
