% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vintages_long}
\alias{vintages_long}
\title{Function to put the vintages data in long format}
\usage{
vintages_long(df, names_to = "pub_date", keep_na = FALSE)
}
\description{
Function to put the vintages data in long format
}
