% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vintages_wide}
\alias{vintages_wide}
\title{Function to put the vintages data to wide format}
\usage{
vintages_wide(df, names_from = "pub_date")
}
\description{
Function to put the vintages data to wide format
}
